# .static/swagger/swagger.yaml
definitions:
...
  api.createTaskRequest:
    properties:
      assignee_id:
        type: integer
      description:
        type: string
      due_date:
        type: string
      priority:
        type: string
      status:
        type: string
      team_id:
        type: integer
      title:
        type: string
    type: object
...
  api.listTaskHistoryResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/api.taskHistoryItemResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  api.listTasksResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/api.taskResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  api.taskHistoryItemResponse:
    properties:
      changed_by:
        type: integer
      created_at:
        type: string
      field_name:
        type: string
      id:
        type: integer
      new_value:
        type: any
      old_value:
        type: any
      task_id:
        type: integer
    type: object
...
  api.taskResponse:
    properties:
      assignee_id:
        type: integer
      created_at:
        type: string
      created_by:
        type: integer
      description:
        type: string
      due_date:
        type: string
      id:
        type: integer
      priority:
        type: string
      status:
        type: string
      team_id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/tasks:
    get:
      parameters:
      - description: Team ID
        in: query
        name: team_id
        required: true
        type: integer
      - description: Status
        in: query
        name: status
        type: string
      - description: Assignee ID
        in: query
        name: assignee_id
        type: integer
      - description: Limit (max 100)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listTasksResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List tasks
      tags:
      - tasks
    post:
      consumes:
      - application/json
      parameters:
      - description: Idempotency key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      - description: Create task
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.createTaskRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create task
      tags:
      - tasks
  /api/v1/tasks/{id}:
    delete:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Idempotency key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete task
      tags:
      - tasks
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.taskResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get task by id
      tags:
      - tasks
    patch:
      consumes:
      - application/json
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Idempotency key for safe retries
        in: header
        name: Idempotency-Key
        type: string
      - description: Patch payload
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Patch task
      tags:
      - tasks
...
  /api/v1/tasks/{id}/history:
    get:
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      - description: Limit (1..100)
        in: query
        name: limit
        type: integer
      - description: Offset (>=0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listTaskHistoryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get task history
      tags:
      - tasks
...
swagger: "2.0"